Move TaskSource to execution package
Extracted-From: https://github.com/prestosql/presto/commit/8ca92f2d4a7fb2ecc912170bf1906494ef061a29
