+++ presto-main/src/main/java/com/facebook/presto/execution/TaskSource.java
@@ -11,10 +11,8 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.facebook.presto;
+package com.facebook.presto.execution;
 
-import com.facebook.presto.execution.Lifespan;
-import com.facebook.presto.execution.ScheduledSplit;
 import com.facebook.presto.sql.planner.plan.PlanNodeId;
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
